test/requirements.txt
streamlit
pillow
deepface
send-mail
pandas